use neo core binary serializer for stack items
